Add validation and examples
I need to make sure we're on a version that supports searching for waits and that the feature is enabled. I also moved the wait search validation up higher to avoid unnecessary work and added the variables to debug output.

There's also a new example query that shows people how to search.