Two for the show
Tinker with entry checks in LogHunter

Bring HumanEvents quick block check code up to date with BlockViewer. This does not change the permanent table code. That's a much larger block of work that I don't have time for right now, because it's not copy/paste, and involved updating the view definitions.