Add searches for forced plans and forced plan failures
This was a bit of an adventure. It started off as me wanting to allow users to search for forced plans, and forced plans that are failing.

Along the way, it made me realize that I had a couple 2017 columns hidden behind 2017 dynamic sql.

It also made me realize that there were a bunch of temp tables (mostly new) that need to be truncated when we're in get all databases mode.

It also made me realize that Microsoft's documentation was incomplete.

It also made me realize that there are some new 2022 columns in query_store_plan that I wasn't including in the results.

A big todo is to audit the docs for other views to see if I'm missing anything.