Reset the value of @query_text_search before each loop, if escaping and hitting multiple databases
This is the smallest change that I could think of that would close #380. I have made an effort to be faithful to the original's style, but I must admit that I have not taken the time to become greatly familiar with this procedure. What I have done appears to work, but I have no idea if this is the best way.