Add beta version of index cleanup
This is the BETA VERSION of a stored procedure I'm working on to automate index cleanup for unused and duplicative indexes.

It requires much testing and probably many fixes before it is complete, or even ready for production use.

There is no harm in running it to help me debug and validate results. It does not take any actions aside from selecting and printing.