dedupe_only mode
trying to add a dedupe_only mode to index cleanup, if someone wants to not trigger a bunch of unused index disable commands and just dedupe existing indexes. this makes sense for servers that aren't up for a long time, and for people who (rightfully) may not trust usage details.